Update MendeleyReferenceManager.download.recipe.yaml
We use %NAME% as a short name of the app and %DISPLAYNAME% as the longer one (with spaces in this case) in downstream recipes so this one breaks on it.  I've suggested adding a AppDmgVersioner processor into this which makes the recipe slightly less fragile if the app name changes (as it detects *.app) and also allows it to run if the %NAME% variable is overridden in downstream recipes.